+++ framework/settings/globalsettings.py
@@ -1,0 +1,1 @@
+mongo_settings_local = 'mongodb://mongodb:27017/'